src/requirements.txt
alembic==1.15.1 ; python_version >= "3.12" and python_version < "4.0"
amqp==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi-bindings==21.2.0 ; python_version >= "3.12" and python_version < "4.0"
argon2-cffi==23.1.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.12" and python_version < "4.0"
celery==5.3.0 ; python_version >= "3.12" and python_version < "4.0"
celery[redis]==5.3.0 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==44.0.2 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cache2[redis]==0.2.2 ; python_version >= "3.12" and python_version < "4.0"
fastapi-users-db-sqlalchemy==7.0.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-users==14.0.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi-users[sqlalchemy]==14.0.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.11 ; python_version >= "3.12" and python_version < "4.0"
flower==2.0.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
humanize==4.12.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
kombu==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.3 ; python_version >= "3.12" and python_version < "4.0"
makefun==1.15.6 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.9 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pendulum==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-client==0.21.1 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.50 ; python_version >= "3.12" and python_version < "4.0"
pwdlib[argon2,bcrypt]==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.8.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.39 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy[asyncio]==2.0.39 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.46.1 ; python_version >= "3.12" and python_version < "4.0"
tornado==6.4.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==1.0.4 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
